+++ src/functionalTest/java/org/zaproxy/gradle/common/spotless/FormatPropertiesStepFunctionalTest.java
@@ -22,34 +22,25 @@ package org.zaproxy.gradle.common.spotless;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.is;
-import static org.hamcrest.Matchers.notNullValue;
 
-import java.io.File;
 import java.io.InputStreamReader;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.List;
 import java.util.Properties;
-import java.util.stream.Collectors;
 import org.gradle.testkit.runner.BuildResult;
-import org.gradle.testkit.runner.BuildTask;
-import org.gradle.testkit.runner.GradleRunner;
-import org.gradle.testkit.runner.TaskOutcome;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.io.TempDir;
 
-public class FormatPropertiesStepFunctionalTest {
+class FormatPropertiesStepFunctionalTest extends FunctionalTest {
     @Test
-    void shouldFormatProperties(@TempDir File testProjectDir) throws Exception {
+    void shouldFormatProperties() throws Exception {
         // Given
         Path originalBuildFilePath =
                 Path.of(getClass().getResource("spotlessProperties.gradle.kts").toURI());
-        Path copiedBuildFilePath = testProjectDir.toPath().resolve("build.gradle.kts");
+        Path copiedBuildFilePath = projectDir.resolve("build.gradle.kts");
         Path originalUnformattedPropertiesPath =
                 Path.of(getClass().getResource("unformatted.properties").toURI());
-        Path copiedUnformattedPropertiesPath =
-                testProjectDir.toPath().resolve("unformatted.properties");
+        Path copiedUnformattedPropertiesPath = projectDir.resolve("unformatted.properties");
         Path formattedPropertiesPath =
                 Path.of(getClass().getResource("formatted.properties").toURI());
         String spotlessPropertiesApplyTaskName = ":spotlessPropertiesApply";
@@ -57,25 +48,12 @@ public class FormatPropertiesStepFunctionalTest {
         Files.copy(originalUnformattedPropertiesPath, copiedUnformattedPropertiesPath);
         Files.copy(originalBuildFilePath, copiedBuildFilePath);
         // When
-        Path pluginClasspath =
-                Path.of(getClass().getClassLoader().getResource("pluginClasspath.txt").toURI());
-        List<File> pluginClasspathFiles =
-                Files.readAllLines(pluginClasspath).stream()
-                        .map(File::new)
-                        .collect(Collectors.toList());
-        BuildResult result =
-                GradleRunner.create()
-                        .withProjectDir(testProjectDir)
-                        .withArguments(spotlessPropertiesApplyTaskName)
-                        .withPluginClasspath(pluginClasspathFiles)
-                        .build();
+        BuildResult result = build(spotlessPropertiesApplyTaskName);
         // Then
-        BuildTask spotlessPropertiesApplyTask = result.task(spotlessPropertiesApplyTaskName);
-        assertThat(spotlessPropertiesApplyTask, is(notNullValue()));
-        assertThat(spotlessPropertiesApplyTask.getOutcome(), is(TaskOutcome.SUCCESS));
+        assertTaskSuccess(result, spotlessPropertiesApplyTaskName);
         assertThat(
-                Files.readString(copiedUnformattedPropertiesPath, StandardCharsets.UTF_8),
-                is(equalTo(Files.readString(formattedPropertiesPath, StandardCharsets.UTF_8))));
+                contentOf(copiedUnformattedPropertiesPath),
+                is(equalTo(contentOf(formattedPropertiesPath))));
 
         var properties = new Properties();
         properties.load(